+++ packages/pwa/app/utils/routes-utils.js
@@ -5,21 +5,20 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-import {getUrlConfig} from './utils'
 import {urlPartPositions} from '../constants'
 
 /**
- * Configure the routes based on url configuration from pwa-kit.config.json file
+ * Creates routes based on the provided application configuration.
  *
  * @param {array} routes - array of routes to be configured
- * @param {object} - a custom configured object
+ * @param {object} - the application configuration object
  * @return {array} - list of configured route objects
  */
-export const configureRoutes = (routes = [], {ignoredRoutes = []}) => {
+export const configureRoutes = (routes = [], config, {ignoredRoutes = []}) => {
     if (!routes.length) return []
+    if (!config) return routes
 
-    const urlConfig = getUrlConfig()
-    if (!urlConfig) return routes
+    const {url: urlConfig} = config?.app
 
     return routes.map((route) => {
         const {path, ...rest} = route